Removed Noto Colour Emoji from body
The way this font is currently being used is making the page somewhat unreadable.

It adds a space between each word that just seems really jarring for no good reason.

"Fedora   Silverblue    is    growing   and    Fedora    in    general   has   a   new strategy.   In   order   to "

Why does it appear that the above text has 3 spaces on my screen?

Without the changes...

"Fedora Silverblue is growing and Fedora in general has a new strategy. In order to "